+++ src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 
@@ -8,6 +9,7 @@ import { CartModule } from './cart/cart.module';
 import { ProductsModule } from './products';
 import { SharedModule } from './shared';
 import { SpinnerModule } from './widgets/spinner/spinner.module';
+import { httpInterceptorProviders } from './core/interceptors';
 
 import { AppComponent } from './app.component';
 
@@ -32,12 +34,14 @@ import { LayoutModule } from './layout/layout.module';
     CartModule,
     LayoutModule,
     SpinnerModule.forRoot(),
+    HttpClientModule,
     AppRoutingModule
   ],
   providers: [
     { provide: LocalStorageService, useClass: LocalStorageService },
     { provide: ConstantsServiceService, useValue: ConstantsServiceServiceValue },
     { provide: RandomString, useFactory: generatorServiceFactory(10) },
+    httpInterceptorProviders,
     ConfigOptionsService
   ],
   bootstrap: [AppComponent]