+++ src/app/core/models/app-settings.model.ts
@@ -1,0 +1,4 @@
+export interface AppSettings {
+    productUrl: string;
+    orderUrl: string;
+}