+++ src/app/core/models/order.model.ts
@@ -1,0 +1,6 @@
+import { CartItem } from '../../cart/models';
+
+export interface Order {
+    id?: number;
+    items: CartItem[];
+}