+++ src/app/core/services/order.service.ts
@@ -2,7 +2,9 @@ import { Injectable } from '@angular/core';
 
 
 import { CartItem } from '../../cart/models';
+import { Order } from '../models/order.model';
 import { LocalStorageService } from './local-storage.service';
+import { OrderHttpClientService } from './order-http-client.service';
 
 @Injectable({
   providedIn: 'root'
@@ -10,36 +12,41 @@ import { LocalStorageService } from './local-storage.service';
 export class OrderService {
 
   constructor(
-    private localStorageService: LocalStorageService
+    private localStorageService: LocalStorageService,
+    private orderHttpClientService: OrderHttpClientService
   ) { }
 
-  addOrder(order: CartItem[]) {
-    const ordersList = this.localStorageService.getItem('orders') || [];
-    if (Array.isArray(ordersList)) {
-      ordersList.push(order);
-    }
-    this.localStorageService.setItem('orders', ordersList);
+  addOrder(items: CartItem[]) {
+    // const ordersList = this.localStorageService.getItem('orders') || [];
+    // if (Array.isArray(ordersList)) {
+    //   ordersList.push(order);
+    // }
+    // this.localStorageService.setItem('orders', ordersList);
+    this.orderHttpClientService.createOrder({ items })
+      .then(order => console.log('order added', order));
   }
 
-  removeOrder(deletingOrder: CartItem[]) {
-    const ordersList = this.getOrderList();
-    let firstMatchIndex = -1;
-    ordersList.forEach((orderItems, index) => {
-      let match = true;
-      deletingOrder.forEach(deletingOrderItem => {
-        if (!orderItems.find(orderItem => orderItem.product.id === deletingOrderItem.product.id)) {
-          match = false;
-        }
-      });
-      if (match && firstMatchIndex === -1) {
-        firstMatchIndex = index;
-      }
-    });
-    ordersList.splice(firstMatchIndex, 1);
-    this.localStorageService.setItem('orders', ordersList);
+  removeOrder(deletingOrder: Order) {
+    return this.orderHttpClientService.deleteOrder(deletingOrder)
+      .then(_ => this.getOrderList().toPromise());
+    // const ordersList = this.getOrderList();
+    // let firstMatchIndex = -1;
+    // ordersList.forEach((orderItems, index) => {
+    //   let match = true;
+    //   deletingOrder.forEach(deletingOrderItem => {
+    //     if (!orderItems.find(orderItem => orderItem.product.id === deletingOrderItem.product.id)) {
+    //       match = false;
+    //     }
+    //   });
+    //   if (match && firstMatchIndex === -1) {
+    //     firstMatchIndex = index;
+    //   }
+    // });
+    // ordersList.splice(firstMatchIndex, 1);
+    // this.localStorageService.setItem('orders', ordersList);
   }
 
-  getOrderList(): CartItem[][] {
-    return this.localStorageService.getItem('orders') || [];
+  getOrderList() {
+    return this.orderHttpClientService.getOrders();
   }
 }